+++ src/pages/AddCard.tsx
@@ -1,5 +1,5 @@
 import CardNumbers from '../components/CardNumbers/CardNumbers';
-import ExpiredDate from '../components/ExpiredDate/ExpiredDate';
+import ExpiredDate from '../components/ExpiredDate/ExpiredDates';
 import CardOwnerName from '../components/CardOwnerName/CardOwnerName';
 import SecurityCode from '../components/SecurityCode/SecurityCode';
 import CardPassword from '../components/CardPassword/CardPassword';
